refactor: 💡 rename all privacy incidences to privacyPools
